move to tns_api
